#537 - Move around some classes, add javadoc, some formatting changes
