+++ recaf-plugin/src/main/java/dev/xdark/recaf/plugin/repository/PluginRepository.java
@@ -1,0 +1,15 @@
+package dev.xdark.recaf.plugin.repository;
+
+import java.util.List;
+
+/**
+ * Remote host for plugins.
+ *
+ * @author xtherk
+ */
+public interface PluginRepository {
+	/**
+	 * @return Plugin information in the remote repo
+	 */
+	List<PluginRepoItem> pluginItems();
+}
