+++ recaf-ui/src/main/java/me/coley/recaf/Main.java
@@ -1,6 +1,6 @@
 package me.coley.recaf;
 
-import dev.xdark.recaf.plugin.RecafPlugin;
+import dev.xdark.recaf.plugin.RecafRootPlugin;
 import me.coley.recaf.launch.InitializerParameters;
 import me.coley.recaf.presentation.PresentationType;
 import me.coley.recaf.scripting.ScriptEngine;
@@ -32,7 +32,8 @@ public class Main {
 	public static void main(String[] args) {
 		setupLogging();
 		// Initialization plugin
-		RecafPlugin.initialize();
+		RecafRootPlugin.initialize();
+		// Read application parameters
 		InitializerParameters parameters = InitializerParameters.fromArgs(args);
 		new Recaf().initialize(parameters);
 		// run script from parameters if found
