+++ recaf-ui/src/main/java/me/coley/recaf/presentation/GuiPresentation.java
@@ -1,6 +1,6 @@
 package me.coley.recaf.presentation;
 
-import dev.xdark.recaf.plugin.RecafPlugin;
+import dev.xdark.recaf.plugin.RecafRootPlugin;
 import javafx.application.Platform;
 import me.coley.recaf.Controller;
 import me.coley.recaf.RecafConstants;
@@ -52,8 +52,13 @@ public class GuiPresentation implements Presentation {
 		// Setup listener to ensure we update classpath dependency directories
 		CompileDependencyUpdater.install(controller);
 		DecompileInterception.install(controller);
-        // For brevity, Move the logic to "PluginLoader"
-		PluginLoader.install();
+		// Enable plugins that are marked as active
+		RecafRootPlugin.getInstance().enablePlugins(Configs.plugin().enabledState
+				.entrySet().stream()
+				.filter(Map.Entry::getValue)
+				.map(Map.Entry::getKey)
+				.collect(Collectors.toSet())
+		);
 		// Open UI
 		JFXUtils.initializePlatform(() -> {
 			try {
@@ -61,7 +66,7 @@ public class GuiPresentation implements Presentation {
 				RecafUI.getWindows().getMainWindow().show();
 			} catch (Throwable ex) {
 				logger.error("Recaf crashed due to an unhandled error." +
-					"Please open a bug report: " + RecafConstants.URL_BUG_REPORT, ex);
+						"Please open a bug report: " + RecafConstants.URL_BUG_REPORT, ex);
 				Platform.exit();
 				System.exit(-1);
 			}
