+++ recaf-ui/src/main/java/me/coley/recaf/ui/control/tree/item/FilterableTreeItem.java
@@ -123,7 +123,7 @@ public class FilterableTreeItem<T> extends TreeItem<T> {
 	 * @param item
 	 * 		Child item to add.
 	 */
-	@SuppressWarnings("unchecked")
+	@SuppressWarnings({"unchecked", "rawtypes"})
 	public void addAndSortChild(TreeItem<T> item) {
 		synchronized (sourceChildren) {
 			int index = Collections.binarySearch((List) getChildren(), item);
