+++ recaf-ui/src/main/java/me/coley/recaf/ui/plugin/MarketplaceTab.java
@@ -1,9 +1,10 @@
 package me.coley.recaf.ui.plugin;
 
+import dev.xdark.recaf.plugin.repository.PluginRepoItem;
 import javafx.scene.control.Tab;
-import me.coley.recaf.ui.plugin.item.MarketplacePluginItem;
-import me.coley.recaf.ui.plugin.repository.CommonPluginRepository;
-import me.coley.recaf.ui.plugin.repository.PluginRepository;
+import me.coley.recaf.ui.plugin.item.RemotePluginItem;
+import dev.xdark.recaf.plugin.repository.CommonPluginRepository;
+import dev.xdark.recaf.plugin.repository.PluginRepository;
 import me.coley.recaf.ui.util.Lang;
 import me.coley.recaf.util.logging.Logging;
 import org.slf4j.Logger;
@@ -23,7 +24,7 @@ public class MarketplaceTab extends Tab {
         super(Lang.get("menu.plugin.marketplace"));
     }
 
-    private List<MarketplacePluginItem> createPluginItems() {
+    private List<PluginRepoItem> createPluginItems() {
         return REPOSITORY.pluginItems();
     }
 }
