+++ recaf-ui/src/main/java/me/coley/recaf/ui/plugin/item/InstalledPluginItem.java
@@ -1,6 +1,7 @@
 package me.coley.recaf.ui.plugin.item;
 
-import dev.xdark.recaf.plugin.RecafPlugin;
+import dev.xdark.recaf.plugin.RecafRootPlugin;
+import dev.xdark.recaf.plugin.repository.PluginRepoItem;
 import me.coley.recaf.config.Configs;
 import me.coley.recaf.util.logging.Logging;
 import org.slf4j.Logger;
@@ -16,64 +17,60 @@ import java.nio.file.Paths;
  *
  * @author xtherk
  */
-public class InstalledPluginItem extends PluginItem {
-    private static final Logger logger = Logging.get(InstalledPluginItem.class);
+public class InstalledPluginItem extends PluginRepoItem {
+	private static final Logger logger = Logging.get(InstalledPluginItem.class);
 
-    /**
-     * @param uri         The path where the plugin is located
-     * @param name        name of the plugin.
-     * @param version     plugin version.
-     * @param author      author of the plugin.
-     * @param description plugin description.
-     */
-    public InstalledPluginItem(URI uri, String name, String version, String author, String description) {
-        super(uri, name, version, author, description);
-    }
+	/**
+	 * @param uri
+	 * 		The path where the plugin is located
+	 * @param name
+	 * 		name of the plugin.
+	 * @param version
+	 * 		plugin version.
+	 * @param author
+	 * 		author of the plugin.
+	 * @param description
+	 * 		plugin description.
+	 */
+	public InstalledPluginItem(URI uri, String name, String version, String author, String description) {
+		super(uri, name, version, author, description);
+	}
 
-    /**
-     * @return Because the Item generated by the installed plugin, it will always return to true
-     */
-    @Override
-    public boolean isInstalled() {
-        return true;
-    }
+	/**
+	 * Disable the plugin
+	 */
+	public void disable() {
+		RecafRootPlugin.getInstance()
+				.getPlugin(name)
+				.ifPresent(pc -> {
+					pc.getLoader().disablePlugin(pc);
+					Configs.plugin().setEnabled(pc.getInformation(), false);
+				});
+	}
 
-    /**
-     * Disable the plugin
-     */
-    public void disable() {
-        RecafPlugin.getInstance()
-                .getPlugin(name)
-                .ifPresent(pc -> {
-                    pc.getLoader().disablePlugin(pc);
-                    Configs.plugin().setEnabled(pc.getInformation(), false);
-                });
-    }
+	/**
+	 * Enable the plugin.
+	 */
+	public void enable() {
+		RecafRootPlugin.getInstance()
+				.getPlugin(name)
+				.ifPresent(pc -> {
+					pc.getLoader().enablePlugin(pc);
+					Configs.plugin().setEnabled(pc.getInformation(), true);
+				});
+	}
 
-    /**
-     * Enable the plugin
-     */
-    public void enable() {
-        RecafPlugin.getInstance()
-                .getPlugin(name)
-                .ifPresent(pc -> {
-                    pc.getLoader().enablePlugin(pc);
-                    Configs.plugin().setEnabled(pc.getInformation(), true);
-                });
-    }
-
-    /**
-     * Uninstall the plugin
-     */
-    public void uninstall() {
-        RecafPlugin.getInstance().unloadPlugin(name);
-        Path path = Paths.get(uri.getPath());
-        try {
-            Files.deleteIfExists(path);
-            Configs.plugin().descriptor.remove(name);
-        } catch (IOException ex) {
-            logger.error("Delete plugin failed", ex);
-        }
-    }
-
+	/**
+	 * Uninstall the plugin.
+	 */
+	public void uninstall() {
+		RecafRootPlugin.getInstance().unloadPlugin(name);
+		Path path = Paths.get(uri.getPath());
+		try {
+			Files.deleteIfExists(path);
+			Configs.plugin().enabledState.remove(name);
+		} catch (IOException ex) {
+			logger.error("Delete plugin failed", ex);
+		}
+	}
 }
