--- recaf-ui/src/main/java/me/coley/recaf/ui/util/HttpUtil.java
@@ -1,63 +1,0 @@
-package me.coley.recaf.ui.util;
-
-import java.io.IOException;
-import java.net.URI;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
-import java.nio.charset.StandardCharsets;
-import java.time.Duration;
-import java.util.concurrent.CompletableFuture;
-
-/**
- * @author xtherk
- */
-public class HttpUtil {
-
-    private static final HttpClient httpClient;
-    private static final int TIMEOUT_SECONDS = 20;
-
-    static {
-        httpClient = HttpClient.newBuilder()
-                .version(HttpClient.Version.HTTP_1_1)
-                .followRedirects(HttpClient.Redirect.NORMAL)
-                .connectTimeout(Duration.ofSeconds(TIMEOUT_SECONDS))
-                .build();
-    }
-
-    public static HttpResponse<byte[]> download(String url) throws IOException, InterruptedException {
-        return download(URI.create(url));
-    }
-
-    public static HttpResponse<byte[]> download(URI uri) throws IOException, InterruptedException {
-        HttpRequest request = HttpRequest.newBuilder().uri(uri).GET().build();
-        return httpClient.send(request, HttpResponse.BodyHandlers.ofByteArray());
-    }
-
-    public static HttpResponse<String> get(String url) throws IOException, InterruptedException {
-        HttpRequest request = HttpRequest.newBuilder().uri(URI.create(url)).GET().build();
-        return httpClient.send(request, HttpResponse.BodyHandlers.ofString(StandardCharsets.UTF_8));
-    }
-
-    public static CompletableFuture<HttpResponse<String>> getAsync(String url) {
-        HttpRequest request = HttpRequest.newBuilder().uri(URI.create(url)).GET().build();
-        return httpClient.sendAsync(request, HttpResponse.BodyHandlers.ofString(StandardCharsets.UTF_8));
-    }
-
-    public static HttpResponse<String> post(String url, String body) throws IOException, InterruptedException {
-        HttpRequest request = HttpRequest.newBuilder()
-                .uri(URI.create(url))
-                .POST(HttpRequest.BodyPublishers.ofString(body))
-                .build();
-        return httpClient.send(request, HttpResponse.BodyHandlers.ofString(StandardCharsets.UTF_8));
-    }
-
-    public static CompletableFuture<HttpResponse<String>> postAsync(String url, String body) {
-        HttpRequest request = HttpRequest.newBuilder()
-                .uri(URI.create(url))
-                .POST(HttpRequest.BodyPublishers.ofString(body))
-                .build();
-        return httpClient.sendAsync(request, HttpResponse.BodyHandlers.ofString(StandardCharsets.UTF_8));
-    }
-
-}
